+++ packages/kiarina/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- **kiarina-utils-file**: Fixed type checking errors when reading comment-only YAML files
+
 ## [1.33.0] - 2026-01-31
 
 ### Added